move featuredCategory into componet/discover/
